Move pages into src
