--- pages/behandling/[behandlingsId].tsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import BehandlingsPage from '../../src/pages/behandling/BehandlingsPage';
-
-export default BehandlingsPage;
