--- pages/person/[...slug].tsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import PersonPage from '../../src/pages/person/PersonPage';
-
-export default PersonPage;
