+++ src/mocks/server.ts
@@ -1,0 +1,4 @@
+import { setupServer } from 'msw/node';
+import { getMockHandlers } from './handlers';
+
+export const server = setupServer(...getMockHandlers());
