+++ src/pages/behandling/[behandlingsId].tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import BehandlingsPage from '../../features/behandling/BehandlingsPage';
+
+export default BehandlingsPage;
