+++ src/pages/person/[...slug].tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import PersonPage from '../../features/person/PersonPage';
+
+export default PersonPage;
